requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
python-dotenv==1.0.0
requests==2.31.0
alpaca-trade-api==3.0.2
newsapi-python==0.2.7
ta==0.10.2
openai==1.2.0
node-cron==3.0.3
node-fetch==2.6.11
node-jsdom==22.1.0
                